:recycle: Implementação dos casos de uso de movies, e vehicles e organização dos packages.
