+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/navigation/MovieDetailsNavigation.kt
@@ -7,8 +7,8 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.SavedStateHandle
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import br.com.gstoduto.starwars.ui.screens.MovieDetailsScreen
-import br.com.gstoduto.starwars.ui.viewmodels.MovieDetailsViewModel
+import br.com.gstoduto.starwars.ui.screens.movie.MovieDetailsScreen
+import br.com.gstoduto.starwars.ui.viewmodels.movie.MovieDetailsViewModel
 import br.com.gstoduto.starwars.util.Constants.ID_MOVIE
 import kotlinx.coroutines.launch
 
@@ -33,12 +33,12 @@ fun NavGraphBuilder.movieDetails(
                 uiState = uiState,
                 onAddToMyListClick = {
                     scope.launch {
-                        viewModel.addToMyList(it)
+                        viewModel.addMovieToMyList(it)
                     }
                 },
                 onRemoveFromMyList = {
                     scope.launch {
-                        viewModel.removeFromMyList(it)
+                        viewModel.removeMovieFromMyList(it)
                     }
                 }
             )
