+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/network/model/vehicle/VehicleResponse.kt
@@ -1,4 +1,4 @@
-package br.com.gstoduto.starwars.network.model
+package br.com.gstoduto.starwars.network.model.vehicle
 
 import br.com.gstoduto.starwars.database.entities.VehicleEntity
