+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/network/services/MovieService.kt
@@ -1,6 +1,6 @@
 package br.com.gstoduto.starwars.network.services
 
-import br.com.gstoduto.starwars.network.model.MoviesResponse
+import br.com.gstoduto.starwars.network.model.movie.MoviesResponse
 import retrofit2.http.GET
 
 interface MovieService {
