+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/network/services/VehicleService.kt
@@ -1,6 +1,6 @@
 package br.com.gstoduto.starwars.network.services
 
-import br.com.gstoduto.starwars.network.model.VehiclesResponse
+import br.com.gstoduto.starwars.network.model.vehicle.VehiclesResponse
 import retrofit2.http.GET
 
 interface VehicleService {
