+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/ui/uistates/movie/MovieDetailsUiState.kt
@@ -1,4 +1,4 @@
-package br.com.gstoduto.starwars.ui.uistates
+package br.com.gstoduto.starwars.ui.uistates.movie
 
 import br.com.gstoduto.starwars.model.Movie
