+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/ui/use_case/movie/AddMovieToMyListUseCase.kt
@@ -1,0 +1,5 @@
+package br.com.gstoduto.starwars.ui.use_case.movie
+
+interface AddMovieToMyListUseCase {
+    suspend fun addMovieToMyList(id: String)
+}
