+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/ui/use_case/movie/GetMoviesUseCase.kt
@@ -1,0 +1,8 @@
+package br.com.gstoduto.starwars.ui.use_case.movie
+
+import br.com.gstoduto.starwars.model.Movie
+import kotlinx.coroutines.flow.Flow
+
+interface GetMoviesUseCase {
+    suspend fun findMovies(): Flow<List<Movie>>
+}
