+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/ui/use_case/movie/GetMyMovieListUseCase.kt
@@ -1,0 +1,8 @@
+package br.com.gstoduto.starwars.ui.use_case.movie
+
+import br.com.gstoduto.starwars.database.entities.MovieEntity
+import kotlinx.coroutines.flow.Flow
+
+interface GetMyMovieListUseCase {
+    suspend fun getMyListMovies(): Flow<List<MovieEntity>>
+}
