+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/ui/use_case/movie/RemoveMovieFromMyListUseCase.kt
@@ -1,0 +1,5 @@
+package br.com.gstoduto.starwars.ui.use_case.movie
+
+interface RemoveMovieFromMyListUseCase {
+    suspend fun removeMovieFromMyList(id: String)
+}
