+++ StarWars/app/src/main/java/br/com/gstoduto/starwars/ui/viewmodels/movie/MovieDetailsViewModel.kt
@@ -1,12 +1,14 @@
-package br.com.gstoduto.starwars.ui.viewmodels
+package br.com.gstoduto.starwars.ui.viewmodels.movie
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import br.com.gstoduto.starwars.database.entities.toMovie
 import br.com.gstoduto.starwars.model.Movie
-import br.com.gstoduto.starwars.repositories.MovieRepository
-import br.com.gstoduto.starwars.ui.uistates.MovieDetailsUiState
+import br.com.gstoduto.starwars.ui.uistates.movie.MovieDetailsUiState
+import br.com.gstoduto.starwars.ui.use_case.movie.AddMovieToMyListUseCaseImpl
+import br.com.gstoduto.starwars.ui.use_case.movie.GetMovieUseCaseImpl
+import br.com.gstoduto.starwars.ui.use_case.movie.RemoveMovieFromMyListUseCaseImpl
 import br.com.gstoduto.starwars.util.Constants.ID_MOVIE
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -18,7 +20,9 @@ import javax.inject.Inject
 @HiltViewModel
 class MovieDetailsViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
-    private val repository: MovieRepository
+    private val getMovieUseCaseImpl: GetMovieUseCaseImpl,
+    private val addMovieToMyListUseCase: AddMovieToMyListUseCaseImpl,
+    private val removeMovieFromMyListUseCase: RemoveMovieFromMyListUseCaseImpl
 ) : ViewModel() {
 
     private val idMovie = savedStateHandle.get<String>(ID_MOVIE)
@@ -35,7 +39,7 @@ class MovieDetailsViewModel @Inject constructor(
 
     private suspend fun loadMovie(title: String) {
         title.let {
-            val movie = repository.findMovie(it)
+            val movie = getMovieUseCaseImpl.findMovie(it)
 
             movie.collect { movieEntity ->
                 _uiState.value = _uiState.value.copy(
@@ -45,11 +49,11 @@ class MovieDetailsViewModel @Inject constructor(
         }
     }
 
-    suspend fun addToMyList(movie: Movie) {
-        repository.addToMyList(movie.title)
+    suspend fun addMovieToMyList(movie: Movie) {
+        addMovieToMyListUseCase.addMovieToMyList(movie.title)
     }
 
-    suspend fun removeFromMyList(movie: Movie) {
-        repository.removeFromMyList(movie.title)
+    suspend fun removeMovieFromMyList(movie: Movie) {
+        removeMovieFromMyListUseCase.removeMovieFromMyList(movie.title)
     }
 }
